Add script for deploying a local setup
The hack/ directory contains scripts and manifests to provision a
metrics+logs stack on a Kind cluster for local testing.

The `hack/setup.sh` script can be run to deploy:
* the kube-prometheus stack for metrics and alerts.
* Loki+promtail for logs.

The `hack/jsonnet` directory contains Jsonnet files using the
kube-prometheus libraries.

The hack/generate.sh script can be used to regenerate the manifests in
hack/manifests after changes to the kube-prometheus version.

Signed-off-by: Simon Pasquier <spasquie@redhat.com>